separate tempinterp unit and integration tests
